#159001: unreproducible NPE.

+++ core.startup/src/org/netbeans/core/startup/ModuleList.java
@@ -1,7 +1,7 @@
 /*
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS HEADER.
  *
- * Copyright 1997-2007 Sun Microsystems, Inc. All rights reserved.
+ * Copyright 1997-2008 Sun Microsystems, Inc. All rights reserved.
  *
  * The contents of this file are subject to the terms of either the GNU
  * General Public License Version 2 only ("GPL") or the Common
@@ -24,7 +24,7 @@
  * Contributor(s):
  *
  * The Original Software is NetBeans. The Initial Developer of the Original
- * Software is Sun Microsystems, Inc. Portions Copyright 1997-2006 Sun
+ * Software is Sun Microsystems, Inc. Portions Copyright 1997-2008 Sun
  * Microsystems, Inc. All Rights Reserved.
  *
  * If you wish your version of this file to be governed by only the CDDL
@@ -44,20 +44,15 @@ package org.netbeans.core.startup;
 import java.beans.PropertyChangeEvent;
 import java.beans.PropertyChangeListener;
 import java.io.BufferedInputStream;
-import java.io.BufferedOutputStream;
 import java.io.ByteArrayInputStream;
 import java.io.ByteArrayOutputStream;
 import java.io.CharArrayWriter;
 import java.io.DataOutputStream;
 import java.io.File;
-import java.io.FileInputStream;
 import java.io.FileNotFoundException;
-import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.InputStream;
-import java.io.ObjectInputStream;
 import java.io.ObjectOutput;
-import java.io.ObjectOutputStream;
 import java.io.OutputStream;
 import java.io.OutputStreamWriter;
 import java.io.Writer;
@@ -92,6 +87,7 @@ import org.openide.modules.Dependency;
 import org.openide.modules.InstalledFileLocator;
 import org.openide.modules.ModuleInstall;
 import org.openide.modules.SpecificationVersion;
+import org.openide.util.Parameters;
 import org.openide.util.RequestProcessor;
 import org.openide.util.Utilities;
 import org.openide.util.WeakSet;
@@ -242,17 +238,20 @@ final class ModuleList implements Stamps.Updater {
                     }
                 }
                 if (! name.equals(props.get("name"))) throw new IOException("Code name mismatch: " /* #25011 */ + name + " vs. " + props.get("name")); // NOI18N
+                Boolean enabledB = (Boolean)props.get("enabled"); // NOI18N
                 String jar = (String)props.get("jar"); // NOI18N
                 File jarFile;
                 try {
                     jarFile = findJarByName(jar, name);
                 } catch (FileNotFoundException fnfe) {
                     //LOG.fine("Cannot find: " + fnfe.getMessage());
-                    ev.log(Events.MISSING_JAR_FILE, new File(fnfe.getMessage()));
-                    try {
-                        f.delete();
-                    } catch (IOException ioe) {
-                        LOG.log(Level.WARNING, null, ioe);
+                    ev.log(Events.MISSING_JAR_FILE, new File(fnfe.getMessage()), enabledB);
+                    if (!Boolean.FALSE.equals(enabledB)) {
+                        try {
+                            f.delete();
+                        } catch (IOException ioe) {
+                            LOG.log(Level.WARNING, null, ioe);
+                        }
                     }
                     continue;
                 }
@@ -264,7 +263,6 @@ final class ModuleList implements Stamps.Updater {
                 history.upgrade(prevRelease, prevSpec);
                 Boolean reloadableB = (Boolean)props.get("reloadable"); // NOI18N
                 boolean reloadable = (reloadableB != null ? reloadableB.booleanValue() : false);
-                Boolean enabledB = (Boolean)props.get("enabled"); // NOI18N
                 boolean enabled = (enabledB != null ? enabledB.booleanValue() : false);
                 Boolean autoloadB = (Boolean)props.get("autoload"); // NOI18N
                 boolean autoload = (autoloadB != null ? autoloadB.booleanValue() : false);
@@ -299,7 +297,7 @@ final class ModuleList implements Stamps.Updater {
                 // Will only really be flushed if mgr props != disk props, i.e
                 // if version changed or could not be enabled.
                 //status.pendingFlush = true;
-                status.diskProps = props;
+                status.setDiskProps(props);
                 statuses.put(name, status);
             } catch (Exception e) {
                 LOG.log(Level.WARNING, "Error encountered while reading " + name, e);
@@ -432,11 +430,12 @@ final class ModuleList implements Stamps.Updater {
         try {
             mgr.enable(modules);
         } catch (InvalidException ie) {
-            LOG.log(Level.WARNING, null, ie);
+            LOG.log(Level.INFO, null, ie);
             Module bad = ie.getModule();
             if (bad == null) throw new IllegalStateException();
-            ev.log(Events.FAILED_INSTALL_NEW_UNEXPECTED, bad, ie);
-            modules.remove(bad);
+            Set<Module> affectedModules = mgr.getModuleInterdependencies (bad, true, true);
+            ev.log(Events.FAILED_INSTALL_NEW_UNEXPECTED, bad, affectedModules, ie);
+            modules.removeAll (affectedModules);
             // Try again without it. Note that some other dependent modules might
             // then be in the missing list for the second round.
             installNew(modules);
@@ -1033,7 +1032,7 @@ final class ModuleList implements Stamps.Updater {
         if (old == null) {
             nue = new DiskStatus();
             nue.module = m;
-            nue.diskProps = computeProperties(m);
+            nue.setDiskProps(computeProperties(m));
         } else {
             nue = old;
         }
@@ -1253,7 +1252,7 @@ final class ModuleList implements Stamps.Updater {
                     LOG.fine("ModuleList: changes are " + changes);
                 }
                 // We need to write changes.
-                status.diskProps = newProps;
+                status.setDiskProps(newProps);
                 try {
                     writeOut(m, status);
                 } catch (IOException ioe) {
@@ -1274,7 +1273,6 @@ final class ModuleList implements Stamps.Updater {
      */
     private Map<String,Object> computeProperties(Module m) {
         if (m.isFixed() || ! m.isValid()) throw new IllegalArgumentException("fixed or invalid: " + m); // NOI18N
-        if (! (m.getHistory() instanceof ModuleHistory)) throw new IllegalArgumentException("weird history: " + m); // NOI18N
         Map<String,Object> p = new HashMap<String,Object>();
         p.put("name", m.getCodeNameBase()); // NOI18N
         int rel = m.getCodeNameRelease();
@@ -1286,16 +1284,18 @@ final class ModuleList implements Stamps.Updater {
             p.put("specversion", spec); // NOI18N
         }
         if (!m.isAutoload() && !m.isEager()) {
-            p.put("enabled", m.isEnabled() ? Boolean.TRUE : Boolean.FALSE); // NOI18N
+            p.put("enabled", m.isEnabled()); // NOI18N
         }
-        p.put("autoload", m.isAutoload() ? Boolean.TRUE : Boolean.FALSE); // NOI18N
-        p.put("eager", m.isEager() ? Boolean.TRUE : Boolean.FALSE); // NOI18N
-        p.put("reloadable", m.isReloadable() ? Boolean.TRUE : Boolean.FALSE); // NOI18N
-        ModuleHistory hist = (ModuleHistory)m.getHistory();
-        p.put("jar", hist.getJar()); // NOI18N
-        if (hist.getInstallerStateChanged()) {
-            p.put("installer", m.getCodeNameBase().replace('.', '-') + ".ser"); // NOI18N
-            p.put("installerState", hist.getInstallerState()); // NOI18N
+        p.put("autoload", m.isAutoload()); // NOI18N
+        p.put("eager", m.isEager()); // NOI18N
+        p.put("reloadable", m.isReloadable()); // NOI18N
+        if (m.getHistory() instanceof ModuleHistory) {
+            ModuleHistory hist = (ModuleHistory) m.getHistory();
+            p.put("jar", hist.getJar()); // NOI18N
+            if (hist.getInstallerStateChanged()) {
+                p.put("installer", m.getCodeNameBase().replace('.', '-') + ".ser"); // NOI18N
+                p.put("installerState", hist.getInstallerState()); // NOI18N
+            }
         }
         return p;
     }
@@ -1619,7 +1619,13 @@ final class ModuleList implements Stamps.Updater {
                     Map<String, Object> props = dirtyprops.get(cnb);
                     if (! cnb.equals(props.get("name"))) throw new IOException("Code name mismatch"); // NOI18N
                     String jar = (String)props.get("jar"); // NOI18N
-                    File jarFile = findJarByName(jar, cnb);
+                    File jarFile;
+                    try {
+                        jarFile = findJarByName(jar, cnb);
+                    } catch (FileNotFoundException fnfe) {
+                        ev.log(Events.MISSING_JAR_FILE, new File(fnfe.getMessage()), true);
+                        continue;
+                    }
                     Boolean reloadableB = (Boolean)props.get("reloadable"); // NOI18N
                     boolean reloadable = (reloadableB != null ? reloadableB.booleanValue() : false);
                     Boolean autoloadB = (Boolean)props.get("autoload"); // NOI18N
@@ -1647,7 +1653,7 @@ final class ModuleList implements Stamps.Updater {
                     DiskStatus status = new DiskStatus();
                     status.module = m;
                     status.file = xmlfile;
-                    status.diskProps = statusProps;
+                    status.setDiskProps(statusProps);
                     statuses.put(cnb, status);
                 }
             }
@@ -1676,7 +1682,10 @@ final class ModuleList implements Stamps.Updater {
                 Map<String, Object> props = entry.getValue();
                 if (props.get("enabled") == null || ! ((Boolean)props.get("enabled")).booleanValue()) { // NOI18N
                     DiskStatus status = statuses.get(cnb);
-                    if (status.diskProps.get("enabled") != null && ((Boolean)status.diskProps.get("enabled")).booleanValue()) { // NOI18N
+                    if (status == null) { // Possible? #159001
+                        continue;
+                    }
+                    if (Boolean.TRUE.equals(status.diskProps.get("enabled"))) { // NOI18N
                         if (! status.module.isEnabled()) throw new IllegalStateException("Already disabled: " + status.module); // NOI18N
                         todisable.add(status.module);
                     }
@@ -1765,7 +1774,7 @@ final class ModuleList implements Stamps.Updater {
                 DiskStatus status = statuses.get(cnb);
                 if (status != null) {
                     Map<String,Object> props = entry.getValue();
-                    status.diskProps = props;
+                    status.setDiskProps(props);
                 }
             }
         }
@@ -1792,6 +1801,10 @@ final class ModuleList implements Stamps.Updater {
         public boolean pendingInstall = false;
         /** properties of the module on disk */
         public Map<String,Object /*String|Integer|Boolean|SpecificationVersion*/> diskProps;
+        void setDiskProps(Map<String,Object> diskProps) {
+            Parameters.notNull("diskProps", diskProps);
+            this.diskProps = diskProps;
+        }
         /** if true, the XML was changed on disk by someone else */
         public boolean dirty = false;
         /** for debugging: */
